requirements.txt
psycopg2-binary==2.8.5  #via PostgreSQL